list를 id 말고 name으로 받아 오는 코드 수정, 오류 확인, 더 있을 수 있으니 확인해 봐야함

+++ src/main/java/com/example/prj2be/controller/ds/DsController.java
@@ -67,7 +67,6 @@ public class DsController {
 
     @GetMapping("name/{name}")
     public Ds map(@PathVariable String name) {
-        System.out.println(name);
         return service.getName(name);
     }
 
@@ -75,13 +74,18 @@ public class DsController {
     public ResponseEntity edit(Ds ds,
                                @RequestParam(value = "updateHolidays[]", required = false) String[] holidays,
                                @RequestParam(value = "uploadFile[]", required = false) MultipartFile[] uploadFile,
-                               @RequestParam(value = "deleteFileIds[]", required = false) List<Integer> deleteFileIds
-                               /*@SessionAttribute(value = "login",required = false) Member login*/) throws IOException {
+                               @RequestParam(value = "deleteFileIds[]", required = false) List<Integer> deleteFileIds,
+                               @SessionAttribute(value = "login",required = false) Member login) throws IOException {
         // 약국 정보 수정
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         // TODO : 멤버 테이블 추가 시 로그인 제약 추가
-//        if (!service.hasAccess(ds.getId(), login)) {
-//            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
-//        }
+        if (!service.hasAccess(ds.getId(), login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         if (service.validate(ds)) {
             if (service.update(ds, uploadFile, deleteFileIds, holidays)) {
                 return ResponseEntity.ok().build();
@@ -94,7 +98,15 @@ public class DsController {
     }
 
     @DeleteMapping("delete/{id}")
-    public ResponseEntity delete(@PathVariable Integer id) {
+    public ResponseEntity delete(@PathVariable Integer id,
+                                 @SessionAttribute(value = "login", required = false)Member login) {
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if (!service.hasAccess(id, login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         if (service.delete(id)) {
             return ResponseEntity.ok().build();
